+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to the "vscode-jq-playground" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.3.0]
+- New commands under *jq playground* group. There are commands to open online manual, tutorial and execute query for active text editor on [jqplay.org](https://jqplay.org/)
+
 ## [0.2.0]
-- Possibility to use workspace files, saved or not with autocomplete filename support for plaintext and json types
+- Possibility to use workspace files, saved or not with autocomplete filename, support plaintext and json types
 
 ## [0.1.0]
 - Fix minor bug